refactor: createArticle
