+++ src/actions/news/create.ts
@@ -3,7 +3,7 @@ import { newsArticleSchema } from './schema'
 import { defaultNewsArticleOldCutoff, newsArticleRealtionsIncluder } from './ConfigVars'
 import prisma from '@/prisma'
 import { readCurrenOmegaOrder } from '@/server/omegaOrder/read'
-import { createArticle } from '@/cms/articles/create'
+import { createArticle } from '@/server/cms/articles/create'
 import { createPrismaActionError, createZodActionError } from '@/actions/error'
 import type { ActionReturn } from '@/actions/Types'
 import type { ExpandedNewsArticle } from './Types'
